fix: pipewire ui fixes and broken controller enumeration
Gamepads were originally enumerated and used if a "Gamepad" type was found.
When support for the DualShock 4 was added, it didn't expose any "Gamepad"
interface, just "Stick". Since some of the other controllers mapped also
expose the "Stick" interface, they would be mapped with those outputs, which
would not contain the same components and so the controller wouldn't map
properly.

This commit updates the enumeration code so the controller enumeration tests
for "Gamepads" first, then if none are found, enumerates "Stick"s.